김경식 : 객체생성 시스템 추가 및 연출 이미지, 값 수정

+++ API_BabaIsYou/GameContent/LogoUI.cpp
@@ -16,5 +16,5 @@ void LogoUI::Start()
 	InitRender("Logo.BMP",
 		GameEngineWindow::GetScreenSize().half(),
 		GameEngineWindow::GetScreenSize(),
-		0, 1, RENDER_ORDER::UIIMAGE, 1);
+		0, 1, static_cast<int>(RENDER_ORDER::UIIMAGE), 1);
 }
